chore: 컨트롤러 분리 및 docs 수정
