+++ src/main/java/hhplus/concert/interfaces/dto/BalanceDto.java
@@ -4,6 +4,11 @@ import lombok.Builder;
 
 public class BalanceDto {
 
+    public record Request (
+            Long amount
+    ) {
+    }
+
     @Builder
     public record Response (
         Long userId,
