+++ src/main/java/hhplus/concert/interfaces/dto/ConcertDto.java
@@ -1,0 +1,21 @@
+package hhplus.concert.interfaces.dto;
+
+import lombok.Builder;
+
+import java.util.List;
+
+public class ConcertDto {
+
+    @Builder
+    public record Response (
+            List<ConcertInfo> concerts
+    ) {
+        @Builder
+        public record ConcertInfo (
+                Long concertId,
+                String title,
+                String description
+        ) {
+        }
+    }
+}
