+++ src/main/java/hhplus/concert/interfaces/dto/PaymentDto.java
@@ -1,10 +1,7 @@
 package hhplus.concert.interfaces.dto;
 
 import hhplus.concert.support.type.PaymentStatus;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 public class PaymentDto {
