+++ src/main/java/hhplus/concert/interfaces/dto/QueueDto.java
@@ -3,7 +3,6 @@ package hhplus.concert.interfaces.dto;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import hhplus.concert.support.type.QueueStatus;
 import lombok.Builder;
-import lombok.Getter;
 
 import java.time.LocalDateTime;
