+++ src/main/java/hhplus/concert/interfaces/dto/ReservationDto.java
@@ -3,7 +3,6 @@ package hhplus.concert.interfaces.dto;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import hhplus.concert.support.type.ReservationStatus;
 import lombok.Builder;
-import lombok.Getter;
 
 import java.time.LocalDateTime;
 import java.util.List;
