+++ src/main/java/hhplus/concert/interfaces/dto/ScheduleDto.java
@@ -2,7 +2,6 @@ package hhplus.concert.interfaces.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Builder;
-import lombok.Getter;
 
 import java.time.LocalDateTime;
 import java.util.List;
