+++ src/main/java/hhplus/concert/interfaces/dto/TokenDto.java
@@ -1,10 +1,7 @@
 package hhplus.concert.interfaces.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 import java.time.LocalDateTime;
