+++ src/main/java/hhplus/concert/support/type/ConcertStatus.java
@@ -1,0 +1,6 @@
+package hhplus.concert.support.type;
+
+public enum ConcertStatus {
+    AVAILABLE,
+    UNAVAILABLE
+}
